- work out mechanics of running scripts where we:
a. can't assume anything is installed
b. no packages can be installed
c. there is a container engine running, but!
   1. we can't assume we can spin up images because podman doesn't make that easy
d. so we have to find existing images that we know have what we need

so we can now:

a. locate docker or podman
b. get an image we can use
c. pipe a bash or python script into the docker/podman run command
d. work around ansible issues with multi-line shell scripts